+++ local_OLD/logger.R
@@ -1,7 +1,7 @@
 #' @title Instance logging
 #'
 #' @description A wrapper for \emph{futile.logger} to log events.
-#' Currently supported logged events: \emph{'start', 'finish', error'}.
+#' Currently supported logged events: \emph{'start', 'finish', 'error'}.
 #'
 #' @param beakr a beakr instance.
 #' @param level the log level (i.e. 'DEBUG', 'INFO', etc.).
@@ -9,10 +9,10 @@
 #' @param echo if TRUE will print the log to console.
 #'
 #' @export
-#' @return A `beakr` App object with added middleware.
+#' @return A \code{Beakr} object with added middleware.
 #' @examples
 #' \dontrun{
-#' new_beakr() %>%
+#' newBeakr() %>%
 #'   http_get('/', function(req, res, err) 'LOG TEST') %>%
 #'   logger() %>%
 #'   listen()
@@ -20,10 +20,16 @@
 #' # DEBUG -datestamp- HTTP | / - GET - REQUEST RECEIVED
 #' # INFO -datestamp- HTTP | / - GET - 200
 #' }
-logger <- function(beakr, level = 'DEBUG', file = NULL, echo = TRUE) {
+
+logger <- function(
+  beakr,
+  level = 'DEBUG',
+  file = NULL,
+  echo = TRUE
+) {
 
   if ( is.null(beakr) ) {
-    beakr <- new_beakr(name = "NULL")
+    beakr <- newBeakr(name = "NULL")
   }
 
   if ( level == 'TRACE' ) {
@@ -52,8 +58,8 @@ logger <- function(beakr, level = 'DEBUG', file = NULL, echo = TRUE) {
   }
 
   try(futile.logger::flog.logger( name = "beakr",
-                              threshold = thresh,
-                              appender = appdr ), silent = TRUE)
+                                  threshold = thresh,
+                                  appender = appdr ), silent = TRUE)
   on(
     beakr = beakr,
     event = 'start',
@@ -104,5 +110,25 @@ logger <- function(beakr, level = 'DEBUG', file = NULL, echo = TRUE) {
 
   return(beakr)
 
+}
+
+#' @export
+#' @title Beakr Event Listener
+#'
+#' @description Add an event listener to a \emph{Beakr} instance. Currently
+#' supported events are \code{"start", "finish", "error"}. The events
+#' \code{"start"} and \code{"finish"} will pass the current state of the
+#' \code{req}, \code{res} and \code{err} objects to the Listener. The
+#' \code{"error"} event will pass string error message.
+#'
+#' @param beakr a beakr instance.
+#' @param event the event to listen for, (\emph{"start", "finish", "error"}).
+#' @param FUN the response middleware function.
+#'
+on <- function(beakr, event, FUN) {
+  l <- Listener$new(event = event, FUN)
+  beakr$router$addListener(l)
+  # .addListener(beakr = beakr, event = event, FUN = FUN)
+  return(beakr)
 }