+++ local_examples/test_repeater_beakr.R
@@ -8,7 +8,7 @@
 library(beakr)
 library(MazamaCoreUtils)
 
-new_beakr() %>%
+newBeakr() %>%
 
   # ----- Welcome --------------------------------------------------------------